+++ reVX/offshore/dist_to_ports_converter.py
@@ -11,7 +11,7 @@ from reV.handlers.exclusions import ExclusionLayers
 
 from reVX.handlers.geotiff import Geotiff
 from reVX.handlers.outputs import Outputs
-from reVX.utilities.exclusions_converter import ExclusionsConverter
+from reVX.utilities import ExclusionsConverter
 
 logger = logging.getLogger(__name__)
